New page content data mapper (PagerSearchContentToDataMapper) for search results and refactor PagerContentToDataMapper
