+++ src/bundle/Controller/SearchController.php
@@ -19,7 +19,7 @@ use EzSystems\EzPlatformAdminUi\Form\Data\Content\Draft\ContentEditData;
 use EzSystems\EzPlatformAdminUi\Form\Data\Search\SearchData;
 use EzSystems\EzPlatformAdminUi\Form\Factory\FormFactory;
 use EzSystems\EzPlatformAdminUi\Form\SubmitHandler;
-use EzSystems\EzPlatformAdminUi\Tab\Dashboard\PagerContentToDataMapper;
+use EzSystems\EzPlatformAdminUi\Search\PagerSearchContentToDataMapper;
 use Pagerfanta\Pagerfanta;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -30,8 +30,8 @@ class SearchController extends Controller
     /** @var \eZ\Publish\API\Repository\SearchService */
     private $searchService;
 
-    /** @var \EzSystems\EzPlatformAdminUi\Tab\Dashboard\PagerContentToDataMapper */
-    private $pagerContentToDataMapper;
+    /** @var \EzSystems\EzPlatformAdminUi\Search\PagerSearchContentToDataMapper */
+    private $pagerSearchContentToDataMapper;
 
     /** @var \Symfony\Component\Routing\Generator\UrlGeneratorInterface */
     private $urlGenerator;
@@ -56,7 +56,7 @@ class SearchController extends Controller
 
     /**
      * @param \eZ\Publish\API\Repository\SearchService $searchService
-     * @param \EzSystems\EzPlatformAdminUi\Tab\Dashboard\PagerContentToDataMapper $pagerContentToDataMapper
+     * @param \EzSystems\EzPlatformAdminUi\Search\PagerContentToDataMapper $pagerSearchContentToDataMapper
      * @param \Symfony\Component\Routing\Generator\UrlGeneratorInterface $urlGenerator
      * @param \EzSystems\EzPlatformAdminUi\Form\Factory\FormFactory $formFactory
      * @param \EzSystems\EzPlatformAdminUi\Form\SubmitHandler $submitHandler
@@ -67,7 +67,7 @@ class SearchController extends Controller
      */
     public function __construct(
         SearchService $searchService,
-        PagerContentToDataMapper $pagerContentToDataMapper,
+        PagerSearchContentToDataMapper $pagerSearchContentToDataMapper,
         UrlGeneratorInterface $urlGenerator,
         FormFactory $formFactory,
         SubmitHandler $submitHandler,
@@ -77,7 +77,7 @@ class SearchController extends Controller
         array $userContentTypeIdentifier
     ) {
         $this->searchService = $searchService;
-        $this->pagerContentToDataMapper = $pagerContentToDataMapper;
+        $this->pagerSearchContentToDataMapper = $pagerSearchContentToDataMapper;
         $this->urlGenerator = $urlGenerator;
         $this->formFactory = $formFactory;
         $this->submitHandler = $submitHandler;
@@ -216,7 +216,7 @@ class SearchController extends Controller
             );
 
             return $this->render('@ezdesign/admin/search/search.html.twig', [
-                'results' => $this->pagerContentToDataMapper->map($pagerfanta),
+                'results' => $this->pagerSearchContentToDataMapper->map($pagerfanta),
                 'form' => $form->createView(),
                 'pager' => $pagerfanta,
                 'form_edit' => $editForm->createView(),
